# openapi.yaml
openapi: 3.1.0
info:
  title: AgentMail API
  version: 1.0.0
  description: >
    API for giving AI agents their own email inboxes — send, receive, and
    manage threaded conversations programmatically.
  x-organization: agentmail-demo
servers:
  - url: https://agentmail-demo.vercel.app
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key (sk_live_...)
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
paths:
...
  /webhooks/inbound:
    post:
      operationId: receiveInboundMail
      summary: Receive inbound mail (provider webhook)
      description: >
        Internal endpoint called by the inbound mail provider (e.g. CloudMailin)
        when external mail arrives at one of your inboxes. Verified via HMAC
        signature in the X-AgentMail-Signature header.
      security: []
      tags: [Internal Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Provider-specific JSON payload.
              additionalProperties: true
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    enum: [true]
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid HMAC signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /webhooks/outbound-status:
    post:
      operationId: receiveOutboundStatus
      summary: Receive delivery status (provider webhook)
      description: >
        Internal endpoint called by the outbound provider with
        delivery/bounce status updates.
      security: []
      tags: [Internal Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                    enum: [true]
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid HMAC signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"